reorganise folder structure to seperate concerns and update README's
